requirements.txt
asgiref==3.8.1
distlib==0.3.8
Django==4.2.15
django-environ==0.11.2
gunicorn==22.0.0
djangorestframework==3.15.2
filelock==3.15.4
platformdirs==4.2.2
psycopg==3.2.1
psycopg2-binary==2.9.9
sqlparse==0.5.1
typing_extensions==4.12.2
virtualenv==20.26.3